pyproject.toml
[project]
name = "chatico-mapper-app"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "lun1tunes",email = "rapt0r3ds@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"

dependencies = [
    "fastapi[standard]>=0.119.1,<0.120.0",
    "uvicorn[standard]>=0.38.0,<0.39.0",
    "sqlalchemy[asyncio]>=2.0.44,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "alembic>=1.17.0,<2.0.0",
    "pydantic-settings>=2.11.0,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "aiohttp>=3.13.1,<4.0.0",
    "aio-pika>=9.3.1,<10.0.0",
    "dependency-injector>=4.41.0,<5.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "httpx>=0.27.0,<1.0.0",
    "redis>=5.0.0,<6.0.0",
    "structlog>=23.0.0,<24.0.0",
    "prometheus-client>=0.20.0,<1.0.0",
    "cryptography>=42.0.0,<43.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "aiosqlite>=0.20.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "pytest-xdist>=3.5.0,<4.0.0",
    "httpx>=0.27.0,<1.0.0",
    "faker>=24.0.0,<25.0.0",
    "factory-boy>=3.3.0,<4.0.0",
    "freezegun>=1.4.0,<2.0.0",
    "pytest-env>=1.1.0,<2.0.0",
    "pytest-clarity>=1.0.0,<2.0.0",
    "pytest-sugar>=0.9.7,<1.0.0",
    "black>=24.0.0,<25.0.0",
    "isort>=5.13.0,<6.0.0",
    "flake8>=7.0.0,<8.0.0",
    "mypy>=1.8.0,<2.0.0",
    "pre-commit>=3.6.0,<4.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
